requirements.txt
fastapi==0.135.3
uvicorn==0.34.0
pydantic==2.12.5
pandas==2.3.3
python-pptx==1.0.2
